add proto,intptrdata to header

+++ DDR_LocalServer/LSClient/LSRemoteClientHeadRuleRouter.cpp
@@ -75,7 +75,7 @@ bool LSRemoteClientHeadRuleRouter::IgnoreBody(std::shared_ptr<BaseSocketContaine
 				auto spClientSession = LSClientManager::Instance()->GetTcpClient()->GetConnectedSession();
 				if (spClientSession)
 				{
-					int IntPtr = (int)(spClientSession.get());
+					long IntPtr = (long)(spClientSession.get());
 					if (passnode.nodetype() == eLocalServer)
 					{
 						if (IntPtr == passnode.receivesessionid())
@@ -92,7 +92,7 @@ bool LSRemoteClientHeadRuleRouter::IgnoreBody(std::shared_ptr<BaseSocketContaine
 
 				for (auto spSessionPair : map)
 				{
-					int IntPtr = (int)(spSessionPair.second.get());
+					long IntPtr = (long)(spSessionPair.second.get());
 					if (passnode.nodetype() == eLocalServer)
 					{
 						if (IntPtr == passnode.receivesessionid())
